requirements.txt
dnspython
pymongo
bs4
discord.py
emoji
pygoogletranslation